requirements.txt
Django==4.0.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.23.1